Improve handling of complex headers' ids
Previously to generate header IDs code relied on parser.IDs interface
implementation:

    type IDs interface {
        Generate(value []byte, kind ast.NodeKind) []byte
        Put(value []byte)
    }

Note how Generate method takes []byte, and not an ast.Node. Now consider the
header with some formatting:

    # Header with *formatting* and a [link](https://example.com/)

For such a header Generate receives []byte holding `Header with *formatting*
and a [link](https://example.com/)` text.

To generate a meaningful ID slug from such a text, we resorted to parse such a
text as a separate markdown document, only extracting the contents of text
nodes: “Header with formatting and a link” for this case.

However, if a header contains a link with the reference (non-inline link), like
in this example document:

    # Header with *formatting* and a [link][1]

    Some text.

    [1]: https://example.com/

Our previous approach fails, because when code parses header value as a
separate markdown document, such one-line document doesn't include the back
reference, and `[link][1]` part is interpreted as plain text.

In order to improve handling of headers with complex formatting, this change
drops usage of parser.IDs interface completely. This interface is generally
useful when there's a need to couple parsing and the following HTML rendering
steps, but for use case of this code we only interested in parsing step.

Instead, track header IDs directly when traversing ast tree, and extract text
directly from *ast.Heading nodes and their possible descendants.